+++ RedPandaIDE/editor.cpp
@@ -263,11 +263,12 @@ void Editor::convertToEncoding(const QByteArray &encoding)
 bool Editor::save(bool force, bool doReparse) {
     if (this->mIsNew && !force) {
         return saveAs();
-    }
+    }    
     while (pMainWindow->parsing()) {
         QThread::msleep(200);
     }
     //is this file writable;
+
     pMainWindow->fileSystemWatcher()->removePath(mFilename);
     try {
         if (pSettings->editor().autoFormatWhenSaved()) {